Updated likes routes
