+++ app/__init__.py
@@ -11,7 +11,7 @@ from .api.user_routes import user_routes
 from .api.auth_routes import auth_routes
 from .api.song_routes import song_routes
 from .api.playlist_routes import playlist_routes
-from .api.like_routes import like_routes
+# from .api.like_routes import like_routes
 from .seeds import seed_commands
 from .config import Config
 
@@ -38,7 +38,7 @@ app.register_blueprint(album_routes, url_prefix='/api/albums')
 # app.register_blueprint(album_song_routes, url_prefix='/api/album_songs')
 app.register_blueprint(song_routes, url_prefix='/api/songs')
 app.register_blueprint(playlist_routes, url_prefix='/api/users/playlists')
-app.register_blueprint(like_routes, url_prefix='/api/likes')
+# app.register_blueprint(like_routes, url_prefix='/api/likes')
 db.init_app(app)
 Migrate(app, db)
