+++ app/api/song_routes.py
@@ -172,3 +172,36 @@ def deleteSong(songId):
     db.session.commit()
 
     return {"message": "Song deleted successfully"}, 200
+
+# Like a Song
+@bp.route('/<int:songId>/likes/', methods=['POST'])
+@login_required
+def likes():
+    """
+    A logged in user can like a song
+    """
+    artistId = current_user.id
+    newLike = Like(artistId, songId)
+    db.session.add(newLike)
+    db.session.commit()
+    return {'message': "Success"}
+
+# Unlike a Song
+@bp.route('/<int:songId>/likes/<int:likeId>', methods=['DELETE'])
+@login_required
+def like(likeId):
+    """
+    A logged in user can unlike a song
+    """
+    like = Like.query.filter_by(id=likeId)
+    
+    if not like:
+        return {"error": {"Like not found"}}, 404
+    
+    if like.artist_id != current_user.id:
+        return {"error": {"You do not have permission to delete this like"}}, 403
+    
+    db.session.delete(like)
+    db.session.commit()
+    
+    return {'message': "Successfully deleted"}, 200
